refactor: create plane/geometry.rs
